[TS SDK] organize file structure
