+++ ecosystem/typescript/sdk/src/aptos_client.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-import { HexString, MaybeHexString } from "./hex_string";
+import { HexString, MaybeHexString } from "./utils/hex_string";
 import {
   clear,
   DEFAULT_TXN_EXP_SEC_FROM_NOW,
@@ -13,7 +13,7 @@ import {
   sleep,
   APTOS_COIN,
 } from "./utils";
-import { AptosAccount } from "./aptos_account";
+import { AptosAccount } from "./plugins";
 import * as Gen from "./generated/index";
 import {
   TxnBuilderTypes,
