+++ ecosystem/typescript/sdk/src/aptos_types/abi.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-import { HexString } from "../hex_string";
+import { HexString } from "../utils/hex_string";
 import { Deserializer, Serializer } from "../bcs";
 import { ScriptABI, EntryFunctionABI, TransactionScriptABI, TypeArgumentABI, ArgumentABI } from "./abi";
 import { TypeTagAddress, TypeTagU64 } from "./type_tag";
