+++ ecosystem/typescript/sdk/src/aptos_types/account_address.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-import { HexString, MaybeHexString } from "../hex_string";
+import { HexString, MaybeHexString } from "../utils/hex_string";
 import { Serializer, Deserializer, Bytes } from "../bcs";
 
 export class AccountAddress {
