+++ ecosystem/typescript/sdk/src/bcs/__tests__/helper.test.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-import { AccountAddress } from "../aptos_types";
-import { Deserializer } from "./deserializer";
+import { AccountAddress } from "../../aptos_types";
+import { Deserializer } from "../deserializer";
 import {
   bcsSerializeBool,
   bcsSerializeBytes,
@@ -17,8 +17,8 @@ import {
   deserializeVector,
   serializeVector,
   serializeVectorWithFunc,
-} from "./helper";
-import { Serializer } from "./serializer";
+} from "../helper";
+import { Serializer } from "../serializer";
 
 test("serializes and deserializes a vector of serializables", () => {
   const address0 = AccountAddress.fromHex("0x1");
