+++ ecosystem/typescript/sdk/src/bcs/__tests__/serializer.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-import { Serializer } from "./serializer";
+import { Serializer } from "../serializer";
 
 describe("BCS Serializer", () => {
   let serializer: Serializer;
