+++ ecosystem/typescript/sdk/src/index.ts
@@ -1,15 +1,12 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-export * from "./aptos_account";
 export * from "./aptos_client";
 export * as BCS from "./bcs";
-export * from "./coin_client";
-export * from "./faucet_client";
-export * from "./hex_string";
-export * from "./token_client";
+export * from "./utils";
 export * from "./transaction_builder";
-export * as TokenTypes from "./token_types";
+export * from "./plugins";
+export * as TokenTypes from "./aptos_types/token_types";
 export * as Types from "./generated/index";
 export { derivePath } from "./utils/hd-key";
 export {
