+++ ecosystem/typescript/sdk/src/plugins/__tests__/coin_client.test.ts
@@ -1,10 +1,9 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-import { AptosClient } from "./aptos_client";
-import { getFaucetClient, longTestTimeout, NODE_URL } from "./utils/test_helper.test";
-import { AptosAccount } from "./aptos_account";
-import { CoinClient } from "./coin_client";
+import { AptosClient } from "../../aptos_client";
+import { getFaucetClient, longTestTimeout, NODE_URL } from "../../utils/__tests__/test_helper.test";
+import { AptosAccount, CoinClient } from "../index";
 
 test(
   "transfer and checkBalance works",
