+++ ecosystem/typescript/sdk/src/plugins/__tests__/faucet_client.test.ts
@@ -1,13 +1,13 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-import { AptosClient } from "./aptos_client";
-import { FaucetClient } from "./faucet_client";
-import { AptosAccount } from "./aptos_account";
-import { HexString } from "./hex_string";
-import * as Gen from "./generated/index";
+import { AptosClient } from "../../aptos_client";
+import { FaucetClient } from "../faucet_client";
+import { AptosAccount } from "..";
+import { HexString } from "../../utils/hex_string";
+import * as Gen from "../../generated/index";
 
-import { NODE_URL, getFaucetClient, longTestTimeout } from "./utils/test_helper.test";
+import { NODE_URL, getFaucetClient, longTestTimeout } from "../../utils/__tests__/test_helper.test";
 
 const aptosCoin = "0x1::coin::CoinStore<0x1::aptos_coin::AptosCoin>";
