+++ ecosystem/typescript/sdk/src/plugins/index.ts
@@ -1,0 +1,3 @@
+export { CoinClient } from "./coin_client";
+export { AptosAccount } from "./aptos_account";
+export { TokenClient } from "./token_client";
