+++ ecosystem/typescript/sdk/src/utils/__tests__/misc.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-import { AptosClient } from "../aptos_client";
+import { AptosClient } from "../../aptos_client";
 
 test("test fixNodeUrl", () => {
   expect(new AptosClient("https://test.com").client.request.config.BASE).toBe("https://test.com/v1");
