+++ ecosystem/typescript/sdk/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./misc";
 export * from "./memoize-decorator";
 export * from "./pagination_helpers";
+export * from "./hex_string";
